requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.18=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
asgiref=3.8.1=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.3.0=pypi_0
build=1.2.2.post1=pypi_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2025.4.26=hbd8a1cb_0
cachetools=5.5.2=pypi_0
certifi=2025.4.26=pypi_0
charset-normalizer=3.4.2=pypi_0
chromadb=1.0.8=pypi_0
click=8.1.8=pypi_0
coloredlogs=15.0.1=pypi_0
deprecated=1.2.18=pypi_0
distro=1.9.0=pypi_0
durationpy=0.9=pypi_0
exceptiongroup=1.2.2=pypi_0
fastapi=0.115.9=pypi_0
filelock=3.18.0=pypi_0
flatbuffers=25.2.10=pypi_0
frozenlist=1.6.0=pypi_0
fsspec=2025.3.2=pypi_0
google-auth=2.40.0=pypi_0
googleapis-common-protos=1.70.0=pypi_0
grpcio=1.67.1=pypi_0
h11=0.16.0=pypi_0
httpcore=1.0.9=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.30.2=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pypi_0
importlib-resources=6.5.2=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.9.0=pypi_0
joblib=1.5.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
kubernetes=32.0.1=pypi_0
ld_impl_linux-64=2.43=h712a8e2_4
libexpat=2.7.0=h5888daf_0
libffi=3.4.6=h2dba641_1
libgcc=14.2.0=h767d61c_2
libgcc-ng=14.2.0=h69a702a_2
libgomp=14.2.0=h767d61c_2
liblzma=5.8.1=hb9d3cd8_0
libnsl=2.0.1=hd590300_0
libsqlite=3.49.1=hee588c1_2
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
litellm=1.68.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
mdurl=0.1.2=pypi_0
milvus-lite=2.4.12=pypi_0
mmh3=5.1.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.4.3=pypi_0
ncurses=6.5=h2d0b736_3
networkx=3.2.1=pypi_0
nltk=3.9.1=pypi_0
numpy=2.0.2=pypi_0
nvidia-cublas-cu12=12.6.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.6.80=pypi_0
nvidia-cuda-nvrtc-cu12=12.6.77=pypi_0
nvidia-cuda-runtime-cu12=12.6.77=pypi_0
nvidia-cudnn-cu12=9.5.1.17=pypi_0
nvidia-cufft-cu12=11.3.0.4=pypi_0
nvidia-cufile-cu12=1.11.1.6=pypi_0
nvidia-curand-cu12=10.3.7.77=pypi_0
nvidia-cusolver-cu12=11.7.1.2=pypi_0
nvidia-cusparse-cu12=12.5.4.2=pypi_0
nvidia-cusparselt-cu12=0.6.3=pypi_0
nvidia-nccl-cu12=2.26.2=pypi_0
nvidia-nvjitlink-cu12=12.6.85=pypi_0
nvidia-nvtx-cu12=12.6.77=pypi_0
oauthlib=3.2.2=pypi_0
ollama=0.4.8=pypi_0
onnxruntime=1.19.2=pypi_0
openai=1.75.0=pypi_0
openssl=3.5.0=h7b32b05_1
opentelemetry-api=1.32.1=pypi_0
opentelemetry-exporter-otlp-proto-common=1.32.1=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.32.1=pypi_0
opentelemetry-instrumentation=0.53b1=pypi_0
opentelemetry-instrumentation-asgi=0.53b1=pypi_0
opentelemetry-instrumentation-fastapi=0.53b1=pypi_0
opentelemetry-proto=1.32.1=pypi_0
opentelemetry-sdk=1.32.1=pypi_0
opentelemetry-semantic-conventions=0.53b1=pypi_0
opentelemetry-util-http=0.53b1=pypi_0
orjson=3.10.18=pypi_0
overrides=7.7.0=pypi_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
pillow=11.2.1=pypi_0
pip=25.1.1=pyh8b19718_0
posthog=4.0.1=pypi_0
propcache=0.3.1=pypi_0
protobuf=5.29.4=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pydantic=2.11.4=pypi_0
pydantic-core=2.33.2=pypi_0
pygments=2.19.1=pypi_0
pymilvus=2.5.8=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
python=3.9.22=h85ef794_1_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.1.0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
rank-bm25=0.2.2=pypi_0
readline=8.2=h8c095d6_2
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rich=14.0.0=pypi_0
rpds-py=0.24.0=pypi_0
rsa=4.9.1=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=pypi_0
sentence-transformers=3.4.0=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=80.1.0=pyhff2d567_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
starlette=0.45.3=pypi_0
sympy=1.14.0=pypi_0
tenacity=9.1.2=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.9.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.20.3=pypi_0
tomli=2.2.1=pypi_0
torch=2.7.0=pypi_0
torchvision=0.22.0=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.46.2=pypi_0
triton=3.3.0=pypi_0
typer=0.15.3=pypi_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.0=pypi_0
tzdata=2025.2=pypi_0
ujson=5.10.0=pypi_0
urllib3=2.4.0=pypi_0
uvicorn=0.34.2=pypi_0
uvloop=0.21.0=pypi_0
watchfiles=1.0.5=pypi_0
websocket-client=1.8.0=pypi_0
websockets=15.0.1=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
wrapt=1.17.2=pypi_0
yarl=1.20.0=pypi_0
zipp=3.21.0=pypi_0
